[AD-213] Review changes. Refactor DocumentDbTableSchemaGenerator.java with helper class. Improve test understanding. https://bitquill.atlassian.net/browse/AD-213
